# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Road",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost/api"
        }
    ],
    "paths": {
        "/tasks": {
            "get": {
                "tags": [
                    "/tasks"
                ],
                "summary": "Display a listing of the resource",
                "description": "get all tasks on database and paginate then",
                "operationId": "App\\Http\\Controllers\\TaskController::index",
                "parameters": [
                    {
                        "name": "only",
                        "in": "query",
                        "description": "filter results using field1;field2;field3...",
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "search results using key:value",
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "operators",
                        "in": "query",
                        "description": "set fileds operators using field1:operator1",
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "order results using field:direction",
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "with",
                        "in": "query",
                        "description": "get relations using relation1;relation2;relation3...",
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "define page",
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "limit per page",
                        "style": "form",
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of tasks"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "/tasks"
                ],
                "summary": "Store a newly created resource in storage.",
                "description": "store a new task on database",
                "operationId": "App\\Http\\Controllers\\TaskController::store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "expired_at": {
                                        "type": "string"
                                    },
                                    "user_id": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "New task created"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/tasks/{task}": {
            "get": {
                "tags": [
                    "/tasks"
                ],
                "summary": "Display the specified resource.",
                "description": "show task",
                "operationId": "App\\Http\\Controllers\\TaskController::show",
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "description": "Task id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show task"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "/tasks"
                ],
                "summary": "Update the specified resource in storage",
                "description": "update a task on database",
                "operationId": "App\\Http\\Controllers\\TaskController::update",
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "description": "Task id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "expired_at": {
                                        "type": "string"
                                    },
                                    "user_id": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Task updated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "/tasks"
                ],
                "summary": "Remove the specified resource from storage.",
                "description": "remove a task on database",
                "operationId": "App\\Http\\Controllers\\TaskController::destroy",
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "description": "Task id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Task deleted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/tokens": {
            "post": {
                "tags": [
                    "/tokens"
                ],
                "summary": "Store a newly created resource in storage.",
                "description": "store a new token on database",
                "operationId": "App\\Http\\Controllers\\TokenController::store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "device": {
                                        "type": "string"
                                    },
                                    "abilities": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "New tokens created"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/tokens/{token}": {
            "delete": {
                "tags": [
                    "/tokens"
                ],
                "summary": "Remove the specified resource from storage.",
                "description": "remove a token on database",
                "operationId": "App\\Http\\Controllers\\TokenController::destroy",
                "parameters": [
                    {
                        "name": "token",
                        "in": "path",
                        "description": "Token id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Token deleted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}